+++ src/Poc.Mobile.App/Views/Credentials/CredentialPage.xaml.cs
@@ -1,0 +1,14 @@
+﻿using Xamarin.Forms;
+using Xamarin.Forms.Xaml;
+
+namespace Poc.Mobile.App.Views.Credentials
+{
+    [XamlCompilation(XamlCompilationOptions.Compile)]
+    public partial class CredentialPage : ContentPage
+    {
+        public CredentialPage()
+        {
+            InitializeComponent();
+        }
+    }
+}